+++ engine/internal/aggregate/controller.go
@@ -65,7 +65,7 @@ func (c *Controller) Handle(
 			Method:         "RouteCommandToInstance",
 			Implementation: c.Config.Handler(),
 			Message:        env.Message,
-			Description:    fmt.Sprintf("routed a command of type %T to an empty instance ID", env.Message),
+			Description:    fmt.Sprintf("routed a command of type %T to an empty ID", env.Message),
 			Location:       location.OfMethod(c.Config.Handler(), "RouteCommandToInstance"),
 		})
 	}