+++ engine/internal/aggregate/controller_test.go
@@ -178,7 +178,7 @@ var _ = g.Describe("type Controller", func() {
 						"Method":         Equal("RouteCommandToInstance"),
 						"Implementation": Equal(config.Handler()),
 						"Message":        Equal(command.Message),
-						"Description":    Equal("routed a command of type fixtures.MessageC to an empty instance ID"),
+						"Description":    Equal("routed a command of type fixtures.MessageC to an empty ID"),
 						"Location": MatchAllFields(
 							Fields{
 								"Func": Not(BeEmpty()),